使用websocket实现在线聊天功能

+++ bee-health-search/src/main/java/io/github/lostblackknight/search/entity/ScheduleESModel.java
@@ -57,7 +57,7 @@ public class ScheduleESModel {
     @Field(type = FieldType.Keyword)
     private String doctorCode;
 
-    @Field(type = FieldType.Integer)
+    @Field(type = FieldType.Long)
     private Long memberId;
 
     /**
